% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{chinook}
\alias{chinook}
\title{Lifespans of microacoustic tags used as part of a juvenile salmon survival study conducted during two seasons.}
\format{
A data frame with 150 rows and 2 variables:
\describe{
\item{days}{life span of acoustic tag, in days}
\item{season}{spring or summer study}
...
}
}
\usage{
chinook
}
\description{
Lifespans of microacoustic tags used as part of a juvenile salmon survival study conducted during two seasons.
}
\keyword{datasets}
