% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fail_pred.R
\name{fail_pred}
\alias{fail_pred}
\title{Failure time predictions}
\usage{
fail_pred(times, pars, model = "gompertz")
}
\arguments{
\item{times}{time vector}

\item{pars}{parameter estimates}

\item{model}{survival model}
}
\value{
failure probability
}
\description{
Geneates predictions from failure time model objects
}
