% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fail_select.R
\name{fail_select}
\alias{fail_select}
\title{Selecting a failure time model from a list}
\usage{
fail_select(mod_ls, model)
}
\arguments{
\item{mod_ls}{failmod_list}

\item{model}{model selected from my list}
}
\value{
failmod_obj that may be used for taglife correction using the "ATLAS" R package.
}
\description{
Select a failure time model from predefined list of candidate models produced by the function "fail_fit()"
}
