% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fc_boot.R
\name{fc_boot}
\alias{fc_boot}
\title{Nonparametric bootstrap of failure time model object}
\usage{
fc_boot(mod_obj, nrep)
}
\arguments{
\item{mod_obj}{failure time model object of class "fc_obj"}

\item{nrep}{number of replicate samples}
}
\value{
matrix of bootstrap parameter estimates with, dimensions (rows= nrep, columns= estimated pars in mod_obj)
}
\description{
Nonparametric bootstrap of failure time model object
}
