% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fc_combine.R
\name{fc_combine}
\alias{fc_combine}
\title{Combination of multiple failure time model objects into a list of models}
\usage{
fc_combine(mod_ls)
}
\arguments{
\item{mod_ls}{list of fc_mods}
}
\value{
fc_list object
}
\description{
Combination of multiple failure time model objects into a list of models
}
\details{
A convenient function for combining model failure time model objecs (fc_obj) into a list object (fc_list).
Model list objects may be ranked according to the Skalski and Whitlock GOF metric using fc_rank
}
