% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fc_diff.R
\name{fc_diff}
\alias{fc_diff}
\title{Log-rank test of two data sets}
\usage{
fc_diff(data, time, group, censorID = NULL)
}
\arguments{
\item{data}{dataframe containing all variables}

\item{time}{numeric failure times}

\item{group}{character or factor grouping variable}

\item{censorID}{logical vector the same length as "time" indicating censored observations}
}
\value{
Returns the results of a log-rank test for comparing two survival distributions.
}
\description{
A log-rank test of two data sets using the "survival" package
}
