% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fc_GOFtest.R
\name{fc_GOFtest}
\alias{fc_GOFtest}
\title{Simulated Kolmogorov-Smirnov Test}
\usage{
fc_GOFtest(times, iters = 50000, model = "gompertz", label = "")
}
\arguments{
\item{times}{numeric vector of failure times}

\item{iters}{replicates for bootstrap (default to 50k)}

\item{label}{optional argument for labeling plots}

\item{dist}{distribution}
}
\value{
P-value and plot of sample distribution of D statistic.
}
\description{
Simulated Kolmogorov-Smirnov Test
}
