% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{fc_mod_ls}
\alias{fc_mod_ls}
\title{List of supported models and parameter names.}
\format{
\describe{named list of default failCompare models}
}
\usage{
fc_mod_ls
}
\description{
List of supported models and parameter names.
}
\keyword{datasets}
