% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fc_plot.R
\name{fc_plot}
\alias{fc_plot}
\title{Ploting failure time and sample survival function}
\usage{
fc_plot(time, surv, group = NULL)
}
\arguments{
\item{time}{}

\item{surv}{}

\item{group}{}
}
\value{
histogram of failure times and scatter plot of sample survival function
}
\description{
Ploting failure time and sample survival function
}
