% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fc_pred.R
\name{fc_pred}
\alias{fc_pred}
\title{Failure time predictions}
\usage{
fc_pred(mod_obj = NULL, times, pars = NULL, model = NULL)
}
\arguments{
\item{times}{time vector}

\item{pars}{parameter estimates}

\item{model}{survival model}
}
\value{
failure probability
}
\description{
Geneates predictions from failure time model objects
}
