% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fc_select.R
\name{fc_select}
\alias{fc_select}
\title{Selecting a failure time model from a list}
\usage{
fc_select(mod_ls, model)
}
\arguments{
\item{mod_ls}{fc_list}

\item{model}{model selected from my list}
}
\value{
fc_obj that may be used for taglife correction using the "ATLAS" R package.
}
\description{
Select a failure time model from predefined list of candidate models produced by the function "fc_fit()"
}
