% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fc_surv.R
\name{fc_surv}
\alias{fc_surv}
\title{Compute sample survival function}
\usage{
fc_surv(time, non_cen = NULL, rt.value = NULL, rc.value = NULL)
}
\arguments{
\item{time}{failure or censoring time}

\item{non_cen}{logical vector the same length as \code{time}, with TRUE indicating censoring time}
}
\value{
sample survival function
}
\description{
Compute sample survival function
}
\details{
Calculates sample survival function accounting for right censoring. In the absence of censoring, uses the basic survival function estimator, otherwise uses the Kaplan-Meier product limit estimate.
}
