% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fc_tryfit.R
\name{fc_tryfit}
\alias{fc_tryfit}
\title{Error handling for fitting failCompare models}
\usage{
fc_tryfit(fit_call, model = "weibull3")
}
\arguments{
\item{fit_call}{call to dependent model fitting functions.}

\item{model}{model argument passed from fc_fit()}
}
\value{
model fitting output
}
\description{
Error handling for fitting failCompare models
}
\details{
Prevents errors from interupting model run
}
