% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_param_nm.R
\name{get_param_nm}
\alias{get_param_nm}
\title{Look-up parameter names for gompertz model}
\usage{
get_param_nm(mod_nm)
}
\arguments{
\item{mod_nm}{character vector of acceptable failCompare model}
}
\value{
dataframe of with model name "model" and "param" names in order reported by optimizer
}
\description{
Look-up parameter names for gompertz model
}
\examples{
# lognormal parameter names 
get_param_nm(mod_nm=c("lognormal"))

# 3-parameter Weibull and Gompertz parameter names 
get_param_nm(mod_nm=c("weibull3","gompertz")) 

}
