% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{goose12}
\alias{goose12}
\title{Lifespans of microacoustic tags uses as part of a juvenile salmon survival studies conducted during two seasons at Little Goose Dam, WA, USA in 2012}
\format{
A data frame with 150 rows and 2 variables:
\describe{
\item{days}{life span of acoustic tag, in days}
\item{season}{spring or summer study}
...
}
}
\usage{
goose12
}
\description{
Lifespans of microacoustic tags uses as part of a juvenile salmon survival studies conducted during two seasons at Little Goose Dam, WA, USA in 2012
}
\keyword{datasets}
