% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ks_boot.R
\name{ks_boot}
\alias{ks_boot}
\title{Simulated Kolmogorov-Smirnov Test}
\usage{
ks_boot(times, reps = 50000, dist = "gompertz", label = "")
}
\arguments{
\item{times}{numeric vector of failure times}

\item{reps}{replicates for bootstrap (default to 50k)}

\item{dist}{distribution}

\item{label}{optional argument for labeling plots}
}
\value{
P-value and plot of sample distribution of D statistic.
}
\description{
Simulated Kolmogorov-Smirnov Test
}
