% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taglife.fn_weib3.R
\name{logweib3.lik}
\alias{logweib3.lik}
\title{3-parameter weibull likelihood}
\usage{
logweib3.lik(x, tags.in)
}
\arguments{
\item{x}{estimated parameters (beta, gamma, eta) or (shape,thrsh,scale)}

\item{tags.in}{observed time to failure}
}
\value{
log likelihood
}
\description{
3-parameter weibull likelihood
}
