% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{pike}
\alias{pike}
\title{Lifespans of rats in an example described in Lee and Wang 2003.}
\format{
A data frame with 35 rows and 2 variables:
\describe{
\item{days}{time until death of fish}
\item{death}{1s indicate observed failures}
...
}
}
\usage{
pike
}
\description{
Lifespans of rats in an example described in Lee and Wang 2003.
}
\keyword{datasets}
