% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.fc_list.R
\name{plot.fc_list}
\alias{plot.fc_list}
\title{Title}
\usage{
\method{plot}{fc_list}(x, model = NULL, type = "data", km = F, res = 100, xlim, ...)
}
\arguments{
\item{x}{failmod_list objects (ranked or not)}

\item{model}{surival model}

\item{km}{Show kaplan-meier estimates}

\item{res}{fineness of survival function preds}

\item{xlim}{x axis limits, used to override default}

\item{...}{ignore}
}
\value{
plot and a message
}
\description{
Title
}
