% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.fc_obj.R
\name{plot.fc_obj}
\alias{plot.fc_obj}
\title{generic function that plots fitted values when a object of class "fc_obj" is called}
\usage{
\method{plot}{fc_obj}(x, km = FALSE, km.ci = FALSE, res = 100, type = "data", ...)
}
\arguments{
\item{x}{of class "fc_obj", created using}

\item{km}{Show kaplan-meier estimates}

\item{km.ci}{Show 95\% confidence limits surrounding kaplan-meier estimates}

\item{res}{Number of evenly space points within the range of the data for plotting}

\item{type}{Plotting survival curve of data ("data") versus difference between Kaplan-Meier estimates and predictions from a parametric model ("resid")}

\item{...}{arguments passed to plot \code{\link{plot}}}
}
\value{
plot and potentially a message about unplotted models in the set
}
\description{
generic function that plots fitted values when a object of class "fc_obj" is called
}
