% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fc_fit.R
\name{print.failmod_list}
\alias{print.failmod_list}
\title{print.failmod_list()}
\usage{
\method{print}{failmod_list}(x, ...)
}
\arguments{
\item{x}{an object of class "failmod_list"}

\item{...}{ignored}
}
\value{
description of list of models
}
\description{
Printed output for class "failmod_list"
}
