% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fc_fit.R
\name{print.failmod_obj}
\alias{print.failmod_obj}
\title{print.failmod_obj()}
\usage{
\method{print}{failmod_obj}(x, ...)
}
\arguments{
\item{x}{an object of class "failmod_obj"}

\item{...}{ignored}
}
\description{
Printed output for class "failmod_obj"
}
