% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fc_fit.R
\name{print.fc_list}
\alias{print.fc_list}
\title{Generic function that limits the amount of output displayed when
an fc_list is called}
\usage{
\method{print}{fc_list}(x, ...)
}
\arguments{
\item{x}{an object of class "fc_list"}

\item{...}{additional arguments to print()}
}
\value{
description of list of models
}
\description{
Printed output for class "fc_list"
}
