% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fc_fit.R
\name{print.fc_obj}
\alias{print.fc_obj}
\title{Generic function that limits the amount of output displayed when
an fc_obj is called}
\usage{
\method{print}{fc_obj}(x, ...)
}
\arguments{
\item{x}{an object of class "fc_obj"}

\item{...}{additional arguments to print()}
}
\description{
Generic function that limits the amount of output displayed when
an fc_obj is called
}
