% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fc_GOFtest.R
\name{pvit09}
\alias{pvit09}
\title{Cumulative distribution function of Vitality 2009 model}
\usage{
pvit09(x, par1, par2, par3, par4)
}
\arguments{
\item{x}{time}

\item{par1}{r}

\item{par2}{s}

\item{par3}{k}

\item{par4}{u}
}
\value{

}
\description{
Cumulative distribution function of Vitality 2009 model
}
