% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fc_test.R
\name{pvit13}
\alias{pvit13}
\title{Cumulative distribution function of Vitality 2013 model}
\usage{
pvit13(x, par1, par2, par3, par4)
}
\arguments{
\item{x}{time}

\item{par1}{r}

\item{par2}{s}

\item{par3}{lambda}

\item{par4}{beta}
}
\value{
cumulative probability
}
\description{
Cumulative distribution function of Vitality 2013 model
}
