% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fc_test.R
\name{rweibull3}
\alias{rweibull3}
\title{random number generation for 3-parameter weibull}
\usage{
rweibull3(n, shape, scale = 1, thres = 0)
}
\arguments{
\item{n}{sample size}

\item{shape}{beta}

\item{scale}{lambda}

\item{thres}{gamma}
}
\value{
vector of random values from the 3-parameter weibull model
}
\description{
random number generation for 3-parameter weibull
}
