% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{steelhead}
\alias{steelhead}
\title{Lifespans of microacoustic tags used as part of a juvenile steelhead survival study with premature tag failures that require censoring at 79 days.}
\format{
A data frame with 82 rows and 1 variables:
\describe{
\item{Day}{life span of acoustic tag, in days}

...
}
}
\usage{
steelhead
}
\description{
Lifespans of microacoustic tags used as part of a juvenile steelhead survival study with premature tag failures that require censoring at 79 days.
}
\keyword{datasets}
