% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fail_fit.R
\name{summary.failmod_list}
\alias{summary.failmod_list}
\title{Sumary of failmod list}
\usage{
\method{summary}{failmod_list}(object, ...)
}
\arguments{
\item{object}{object of class failmod_list}

\item{...}{ignore}
}
\value{
Summary of model fitting calls and GOF rankings (if available)
}
\description{
Sumary of failmod list
}
\details{
Detailed summary of model objects in failmod list object
}
