% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fail_fit.R
\name{summary.failmod_obj}
\alias{summary.failmod_obj}
\title{Summary of failmod object}
\usage{
\method{summary}{failmod_obj}(object, ...)
}
\arguments{
\item{object}{object of class failmod_obj}

\item{...}{ignore}
}
\value{
Summary of failmod_obj model fit calls
}
\description{
Summary of failmod object
}
\details{
Detailed summary of model objects in failmod object
}
