% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fc_fit.R
\name{summary.fc_list}
\alias{summary.fc_list}
\title{Generic function for summarizing an object of class "fc_list"}
\usage{
\method{summary}{fc_list}(object, ...)
}
\arguments{
\item{object}{object of class fc_list}

\item{...}{additional arguments to summary}
}
\value{
Summary of model fitting calls and GOF rankings (if available)
}
\description{
Generic function for summarizing an object of class "fc_list"
}
