% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fc_fit.R
\name{summary.fc_obj}
\alias{summary.fc_obj}
\title{Generic function for summarizing an object of class "fc_obj"}
\usage{
\method{summary}{fc_obj}(object, ...)
}
\arguments{
\item{object}{object of class fc_obj}

\item{...}{additional arguments to summary}
}
\value{
Summary of fc_obj model of calls to model fitting functions.
}
\description{
Generic function for summarizing an object of class "fc_obj"
}
