% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{trout}
\alias{trout}
\title{Lifespans of rainbow trout exposed to gas supersaturation. Based on example given in Salinger et al. 2003.}
\format{
A data frame with 35 rows and 2 variables:
\describe{
\item{days}{time until death of fish}
\item{censored}{Observation was censored}
...
}
}
\usage{
trout
}
\description{
Lifespans of rainbow trout exposed to gas supersaturation. Based on example given in Salinger et al. 2003.
}
\keyword{datasets}
