% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{chinook}
\alias{chinook}
\title{Lifespans of microacoustic tags uses as part of a juvenile salmon survival studies conducted during two seasons}
\format{
A data frame with 150 rows and 2 variables:
\describe{
\item{days}{life span of acoustic tag, in days}
\item{season}{spring or summer study}
...
}
}
\usage{
chinook
}
\description{
Lifespans of microacoustic tags uses as part of a juvenile salmon survival studies conducted during two seasons
}
\keyword{datasets}
