% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fc_combine.R
\name{fc_combine}
\alias{fc_combine}
\title{Combination of multiple failure time model objects into a list of models}
\usage{
fc_combine(mod_ls)
}
\arguments{
\item{mod_ls}{list of fc_mods}
}
\value{
fc_list object
}
\description{
Combination of multiple failure time model objects into a list of models
}
\details{
A convenience function for combining model failure time model objecs \code{fc_obj} into a failure model
list object \code{fc_list}. Lists that include the "kaplan-meier" model or duplicates are not allowed.
}
\references{
Li, T., and Anderson, J.J. 2009. The vitality model: A way to understand population survival and demographic heterogeneity. Theoretical Population Biology 76(2): 118–131.

Li, T., and Anderson, J.J. 2013. Shaping human mortality patterns through intrinsic and extrinsic vitality processes. Demographic research 28: 341–372.
}
\seealso{
\code{fc_select} and \code{fc_fit}
}
