% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fc_combine.R
\name{fc_combine}
\alias{fc_combine}
\title{Combination of multiple failure time model objects into a list of models}
\usage{
fc_combine(mod_ls)
}
\arguments{
\item{mod_ls}{list of fc_mods}
}
\value{
fc_list object that may be used by
}
\description{
Combination of multiple failure time model objects into a list of models
}
\details{
A convenience function for combining model failure time model objects into a failure model
list object \code{fc_list}. Lists that include the "kaplan-meier" model or duplicates are not allowed.
}
\examples{

### Load example dataframe
data(sockeye)
taglife=sockeye[,"days"] #define vector of times

### Fit a 2-parameter Weibull model
weib_mod=fc_fit(time=taglife,model="weibull")

### Fit a 4-parameter Vitality 2013 model
vit_mod=fc_fit(time=taglife,model="vitality.4p")

# Combine two "fc_obj" objects into a model list of class "fc_list" 
fc_combine(mod_ls = list(weib_mod,vit_mod))

}
\references{
Li, T., and Anderson, J.J. 2009. The vitality model: A way to understand population survival and demographic heterogeneity. Theoretical Population Biology 76(2): 118–131.

Li, T., and Anderson, J.J. 2013. Shaping human mortality patterns through intrinsic and extrinsic vitality processes. Demographic research 28: 341–372.
}
\seealso{
\code{\link{fc_select}} \code{\link{fc_fit}}
}
