% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fc_diff.R
\name{fc_diff}
\alias{fc_diff}
\title{Log-rank test of two data sets}
\usage{
fc_diff(data, time, group, censorID = NULL)
}
\arguments{
\item{data}{dataframe containing all variables}

\item{time}{failure times}

\item{group}{grouping variable}

\item{censorID}{censored observation vector (T/F)}
}
\value{
Results of a log-rank test for comparing two survival distributions
}
\description{
Log-rank test of two data sets
}
