% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fc_fit.R
\name{fc_fit}
\alias{fc_fit}
\title{Fitting one or a set of failure time models}
\usage{
fc_fit(time, model, censorID = NULL, rc.value = NULL, ...)
}
\arguments{
\item{time}{numeric vector of failure times}

\item{model}{character vector of specified model(s)}

\item{censorID}{censored observation vector (T/F)}

\item{rc.value}{right-censoring cutoff value; only observations with times > rc.value are considered to have failed prematurely.}

\item{...}{additional arguments passed}
}
\value{
Returns failure model object of class \code{"fc_obj"} if one model specified OR
a failure model list object of class \code{"fc_list"} if multiple models are specified
}
\description{
Code for fitting a particular tag-life model or models
}
\details{
Model fitting routine used to fit one or a set of failure time models.

The following nine model options include
\code{weibull} = 2-parameter Weibull
\code{weibull3} = 3-parameter Weibull
\code{gompertz} = Gompertz
\code{lognormal} = lognormal
\code{llogis} = log-logistic
\code{gamma} = gamma (2-parameter)
\code{gengamma} = generalized gamma (3-parameter)
\code{vitality.ku} = vitality (2009)
\code{vitality.4p} = vitality (2013)
\code{kaplan-meier} = Kaplan-Meier nonparametric estimate (NOTE: this model cannot be specified in a list with any other model)

If a single model is specified a "fc_obj" is created, which can be
used to adjust a CJS model in the "ATLAS" package.

If multiple models are specified a "fc_list" is created containing
output from all model fits. This object serves as an input in the
fc_rank() function, which ranks the performance of the model using
the \href{http://animalbiotelemetry.biomedcentral.com/articles/10.1186/s40317-020-00213-z}{Skalski and Whitlock (2020)} GOF measure.

Parameter estimates and accompanying standard errors(if available) are printed.
The \code{"fc_obJ"} is a list of the following extractable objects:
\itemize{
\item"mod_choice"  = official model name
\item     "times"  = dataframe of failure time, survival fraction, and censoring binary var.
\item  "fit_vals"  = failure times and predicted survival under the model, 95\% LCL an UCL if available
\item  "mod_objs"  = actual model object created by "flexsurvdist" or "vitality package"-- much more to extract from "flexsurvdist
\item   "par_tab"  = table of parameter estimates and SE in failCompare recognized order
\item     "KM_DF"  = table of K-M estimates for plotting
\item    "KM_mod"  = survival package K-M model estimates
\item  'censored'  = T/F of censored dataset or not
}
}
\references{
Skalski, J. R., and S. L. Whitlock. 2020. Vitality models found useful in modeling tag-failure times in acoustic-tag survival studies. Animal Biotelemetry 8:1–10.DOI:10.1186/s40317-020-00213-z
}
