% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fc_fit.R
\name{fc_fit}
\alias{fc_fit}
\title{Fitting one or a set of failure time models}
\usage{
fc_fit(time, model, rc.value = NULL, rt.value = NULL, ...)
}
\arguments{
\item{time}{numeric vector of failure times}

\item{model}{character vector of specified model(s)}

\item{rc.value}{right-censoring cutoff value; only observations with times > rc.value are considered to have failed prematurely.}

\item{rt.value}{right-truncation cutoff value; only observations with times < rt.value will be included in the model fitting}

\item{...}{additional arguments passed}
}
\value{
A failure model object, if one model specified OR a failure model list object if multiple models are specified
}
\description{
Code for fitting a particular tag-life model or models
}
\details{
Model fitting routine used to fit one or a set of failure time models

If a single model is specified a "fc_obj" is created, which can be
used to adjust a CJS model in the "ATLAS" package.

If multiple models are specified a "fc_list" is created containing
output from all model fits. This object serves as an input in the
fail_rank() function, which ranks the performance of the model using
the \href{http://animalbiotelemetry.biomedcentral.com/articles/10.1186/s40317-020-00213-z}{Skalski and Whitlock (2020)} GOF measure.

(named list)
"mod_choice" = character object of model name or vector of model names,
"fit_vals" = data frame with fitted values for each time and 95\% confidence interval
"mod_objs" = model objects from various packages: kaplan-meiser ("surival"), vitality models ("vitality")
"par_tab" = dataframe of one or more model's parameter estimates
}
