% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fc_fit_single.R
\name{fc_fit_single}
\alias{fc_fit_single}
\title{Fitting a single failure time model}
\usage{
fc_fit_single(time, model, SEs = TRUE, censorID = NULL, rc.value = NULL, ...)
}
\arguments{
\item{time}{failure time}

\item{model}{failure time model}

\item{SEs}{calculating standard errors}

\item{censorID}{logical whether an observation is non-censored}

\item{rc.value}{value after which}

\item{...}{additional arguments passed to fc_tryfit()}
}
\value{

}
\description{
Fitting a single failure time model
}
\details{
function for fitting an indiviudal failure time model assuming inputs have been vetted by primary function fc_fit().
}
