% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fc_fit_single.R
\name{fc_fit_single}
\alias{fc_fit_single}
\title{Fitting a single failure time model}
\usage{
fc_fit_single(y, y_sfrac, model, Hess, non_cen, KM_DF, KM_mod, inits, ...)
}
\arguments{
\item{y}{failure time}

\item{y_sfrac}{survival fraction}

\item{model}{failure time model}

\item{Hess}{calculating standard errors}

\item{non_cen}{logical of length(y)}

\item{KM_DF}{K-M model predictions}

\item{KM_mod}{K-M model object}

\item{inits}{initial value arguments}

\item{...}{additional arguments passed to optimizer}
}
\value{
"fc_obj" if successful NULL if otherwise
}
\description{
Fitting a single failure time model
}
\details{
function for fitting an indiviudal failure time model assuming inputs have been vetted by user-facing function fc_fit().
}
