% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fc_fit_single.R
\name{fc_fit_single}
\alias{fc_fit_single}
\title{Fitting a single failure time model}
\usage{
fc_fit_single(y, y_sfrac, model, Hess, non_cen, KM_DF, KM_mod, ...)
}
\arguments{
\item{y}{failure time}

\item{model}{failure time model}

\item{Hess}{calculating standard errors}

\item{non_cen}{logical of length(y)}

\item{...}{additional arguments passed to fc_tryfit()}
}
\value{

}
\description{
Fitting a single failure time model
}
\details{
function for fitting an indiviudal failure time model assuming inputs have been vetted by primary function fc_fit().
}
