% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fc_rank.R
\name{fc_rank}
\alias{fc_rank}
\title{Ranking failure time models based on goodness of fit (GOF)}
\usage{
fc_rank(x)
}
\arguments{
\item{x}{an object of class "fc_list"}
}
\value{
Table of models ranked according to GOF measure
}
\description{
Ranking failure time models based on goodness of fit (GOF)
}
\details{
Goodness of fit measure from Skalski and Whitlock (2020)
}
