% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fc_select.R
\name{fc_select}
\alias{fc_select}
\title{Selecting a failure time model from a list}
\usage{
fc_select(mod_ls, model)
}
\arguments{
\item{mod_ls}{failure model list object (i.e., class \code{fc_list})}

\item{model}{model selected from list of those available. Possible options include:
\itemize{
\item  "weibull"  = 2-parameter Weibull
\item  "weibull3" = 3-parameter Weibull
\item  "gompertz"  = Gompertz Model
\item  "gamma"  = Gamma distribution (2-parameter)
\item "lognormal"  = Log-Normal distribution
\item  "llogis" = Log-Logistic distribution
\item  "gengamma"  = Generalized Gamma Distribution (3-parameter; Prentice 1974 parameterization)
\item  'vitality.ku'  = 4-parameter vitality model from Li and Anderson (2009)
\item  'vitality.4p'  = 4-parameter vitality model from Li and Anderson (2013)
}}
}
\value{
Returns a failure time model object of class \code{fc_obj} .
}
\description{
This selects a failure time model from predefined list of candidate models produced by the function \code{fc_fit()}
}
