% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fc_surv.R
\name{fc_surv}
\alias{fc_surv}
\title{Compute sample survival function}
\usage{
fc_surv(time, censorID = NULL, rc.value = NULL)
}
\arguments{
\item{time}{failure or censoring time}

\item{censorID}{logical vector the same length as \code{time}, with TRUE indicating censoring time}

\item{rc.value}{time after which all values are censored}
}
\value{
This returns a numeric vector of survival fraction estimates sample survival function.
}
\description{
This computes a sample survival function.
}
\details{
This calculates a sample survival function accounting for right censoring. In the absence of censoring, it uses the basic survival function estimator, or otherwise uses the Kaplan-Meier product limit estimate.
}
