% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fc_test.R
\name{fc_test}
\alias{fc_test}
\title{Log-rank test of two data sets}
\usage{
fc_test(data, time, group, non_cen = NULL)
}
\arguments{
\item{data}{}

\item{time}{}

\item{group}{}
}
\value{

}
\description{
Log-rank test of two data sets
}
