% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fc_test.R
\name{fc_test}
\alias{fc_test}
\title{Simulated Kolmogorov-Smirnov Test}
\usage{
fc_test(times, iters = 50000, model = "gompertz", label = "", plot = FALSE)
}
\arguments{
\item{times}{numeric vector of failure times}

\item{iters}{replicates for bootstrap (default to 50k)}

\item{model}{distribution}

\item{label}{optional argument for labeling plots}

\item{plot}{optional argument for creating histogram}
}
\value{
Returns a P-value and histogram of sample distribution of D statistic.
}
\description{
A simulated Kolmogorov-Smirnov test.
}
\details{
Performs a a simulation-based Kolmogorov-Smirnov test.
}
\seealso{
\code{\link[stats]{ks.test}}.
\code{\link[stats]{rweibull}}.
}
