% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fc_test.R
\name{ks.test_fc}
\alias{ks.test_fc}
\title{ks.test with suppressed warnings}
\usage{
ks.test_fc(...)
}
\arguments{
\item{...}{inputs to stats::ks.test() function}
}
\value{
expected output from ks.test
}
\description{
ks.test with suppressed warnings
}
