% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nll_vitality_ku.R
\name{nll_vitality_ku}
\alias{nll_vitality_ku}
\title{Title}
\usage{
nll_vitality_ku(params, data_time, non_cen, indiv_nll = FALSE)
}
\arguments{
\item{params}{vitality parameters (r,s,k,u)}

\item{data_time}{failure times}

\item{non_cen}{binary indicator variable for non-censored obs (non-censored=1; censored=0)}

\item{indiv_nll}{logical; indicates whether individual likelihood contribution (defaults to FALSE)}
}
\value{
negative log likelihood based on individual time
}
\description{
Title
}
\examples{

data("sockeye")
vit_mod=failCompare::fc_fit(time=sockeye$days,model="vitality.ku")

nll_vitality_ku(params = vit_mod$par_tab[, "params"],
                data_time = vit_mod$times$time,
                non_cen = vit_mod$times$non_cen)

nll_vitality_ku(params = vit_mod$par_tab[, "params"],
                data_time = vit_mod$times$time,
                non_cen = vit_mod$times$non_cen,
                indiv_nll=TRUE)


}
