% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.fc_list.R
\name{plot.fc_list}
\alias{plot.fc_list}
\title{Title}
\usage{
\method{plot}{fc_list}(x, model = NULL, type = "data", km = F, res = 100, xlim, ...)
}
\arguments{
\item{x}{failmod_list objects (ranked or not)}

\item{model}{survival model}

\item{type}{"data" or "residual" plot.}

\item{km}{show kaplan-meier estimates}

\item{res}{fineness of survival function preds}

\item{xlim}{x axis limits, used to override default}

\item{...}{ignore}
}
\value{
plot and a message
}
\description{
Title
}
\details{
Plot type "data" shown by default. For "residual" type plot showing (kaplan-meier estimates - parametric model fit) plot a singual model of class=fc_obj.
}
