% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fc_fit.R
\name{print.fc_list}
\alias{print.fc_list}
\title{print.fc_list()}
\usage{
\method{print}{fc_list}(x, ...)
}
\arguments{
\item{x}{an object of class "fc_list"}

\item{...}{ignored}
}
\value{
description of list of models
}
\description{
Printed output for class "fc_list"
}
