% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fc_fit.R
\name{print.fc_obj}
\alias{print.fc_obj}
\title{print.fc_obj()}
\usage{
\method{print}{fc_obj}(x, ...)
}
\arguments{
\item{x}{an object of class "fc_obj"}

\item{...}{ignored}
}
\description{
Printed output for class "fc_obj"
}
