% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fc_test.R
\name{rvitality}
\alias{rvitality}
\title{Generating  samples from 2009 and 2013 Vitality models}
\usage{
rvitality(
  parms,
  times_dat,
  t_seq_fineness = 0.005,
  quant_seq = seq(0, 1, 0.005),
  model = "Vitality09"
)
}
\arguments{
\item{parms}{vector of parameters, Vitality 2009 (r,s,k,u), Vitality 2013 (r,s,lambda,beta)}

\item{times_dat}{survival times used for determining # samples to generate and range of slices}

\item{t_seq_fineness}{time increments to with which to slice up the survival curve}

\item{quant_seq}{bins in which to place simulated times}

\item{model}{either "Vitality09" ot "Vitality13"}
}
\value{

}
\description{
Generating  samples from 2009 and 2013 Vitality models
}
