% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{sockeye}
\alias{sockeye}
\title{Lifespans of microacoustic tags uses as part of a juvenile salmon survival study conducted at Rocky Reach Dam, WA, USA in 2004}
\format{
A data frame with 50 rows and 1 variable:
\describe{
\item{days}{life span of acoustic tag, in days}
...
}
}
\usage{
sockeye
}
\description{
Lifespans of microacoustic tags uses as part of a juvenile salmon survival study conducted at Rocky Reach Dam, WA, USA in 2004
}
\keyword{datasets}
