% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fc_fit.R
\name{summary.fc_list}
\alias{summary.fc_list}
\title{Sumary of fc list}
\usage{
\method{summary}{fc_list}(object, ...)
}
\arguments{
\item{object}{object of class fc_list}

\item{...}{ignore}
}
\value{
Summary of model fitting calls and GOF rankings (if available)
}
\description{
Sumary of fc list
}
\details{
Detailed summary of model objects in fc list object
}
