% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fc_fit.R
\name{summary.fc_obj}
\alias{summary.fc_obj}
\title{Summary of fc object}
\usage{
\method{summary}{fc_obj}(object, ...)
}
\arguments{
\item{object}{object of class fc_obj}

\item{...}{ignore}
}
\value{
Summary of fc_obj model fit calls
}
\description{
Summary of fc object
}
\details{
Detailed summary of model objects in fc object
}
